move to read-only sniff
