--- packages/CodingStandard/tests/Fixer/Strict/NoUnusedPublicMethodFixer/Test.php
@@ -1,34 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\CodingStandard\Tests\Fixer\Strict\NoUnusedPublicMethodFixer;
-
-use PhpCsFixer\Fixer\FixerInterface;
-use Symplify\CodingStandard\Fixer\Strict\NoUnusedPublicMethodFixer;
-use Symplify\EasyCodingStandard\Tests\AbstractDualRunTestCase;
-
-final class Test extends AbstractDualRunTestCase
-{
-    /**
-     * @dataProvider provideFixCases()
-     */
-    public function testFix(string $expected, ?string $input = null): void
-    {
-        $this->doTest($expected, $input);
-    }
-
-    /**
-     * @return string[][]
-     */
-    public function provideFixCases(): array
-    {
-        return [
-            # wrong => fixed
-            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc'],
-        ];
-    }
-
-    protected function createFixer(): FixerInterface
-    {
-        return new NoUnusedPublicMethodFixer();
-    }
-}
