--- packages/CodingStandard/tests/Fixer/Strict/NoUnusedPublicMethodFixer/fixed/fixed.php.inc
@@ -1,16 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-
-class SomeClass
-{
-
-    public function anotherMethod()
-    {
-
-    }
-}
-
-$someObject = new SomeClass;
-$someObject->anotherMethod();
